requirements.txt
# ============================================================================
# SMARTLOCK BACKEND - COMPATIBLE REQUIREMENTS
# Python 3.11 or 3.12
# ============================================================================

# ----------------------------------------------------------------------------
# DJANGO CORE
# ----------------------------------------------------------------------------
Django==4.2.16
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.5

# ----------------------------------------------------------------------------
# CACHING & TASK QUEUE
# ----------------------------------------------------------------------------
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ----------------------------------------------------------------------------
# AUTHENTICATION & SECURITY
# ----------------------------------------------------------------------------
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
cryptography==41.0.7
django-axes==6.1.1
argon2-cffi==23.1.0

# ----------------------------------------------------------------------------
# MQTT
# ----------------------------------------------------------------------------
paho-mqtt==1.6.1

# ----------------------------------------------------------------------------
# MONITORING & LOGGING
# ----------------------------------------------------------------------------
sentry-sdk==1.39.2
django-health-check==3.18.1

# ----------------------------------------------------------------------------
# API DOCUMENTATION
# ----------------------------------------------------------------------------
drf-spectacular==0.27.0

# ----------------------------------------------------------------------------
# UTILITIES
# ----------------------------------------------------------------------------
python-dotenv==1.0.0
pytz==2023.3.post1
python-dateutil==2.8.2
requests==2.31.0

# ----------------------------------------------------------------------------
# STATIC FILES
# ----------------------------------------------------------------------------
whitenoise==6.6.0

# ----------------------------------------------------------------------------
# PRODUCTION SERVER & DATABASE
# ----------------------------------------------------------------------------
gunicorn==21.2.0
psycopg2-binary==2.9.9

# ----------------------------------------------------------------------------
# HEALTH CHECKS
# ----------------------------------------------------------------------------
django-health-check[psutil]==3.18.1